Suppress conversion warning in Ratpack/conv.cpp.
When building for x64, the compiler complains:
    warning C4267: 'argument': conversion from 'size_t' to 'ULONG', possible loss of data

In practice, the number string will not exceed a ULONG in length.

+++ src/CalcManager/Ratpack/conv.cpp
@@ -578,7 +578,7 @@ PNUMBER StringToNumber(wstring_view numberString, uint32_t radix, int32_t precis
     long expValue = 0L;          // expValue is exponent mantissa, should be unsigned
 
     PNUMBER pnumret = nullptr;
-    createnum(pnumret, numberString.length());
+    createnum(pnumret, static_cast<ULONG>(numberString.length()));
     pnumret->sign = 1L;
     pnumret->cdigit = 0;
     pnumret->exp = 0;
